Logger: Implement FunctionOrigin option, to substitute logger name for function's
This follows the well-established practice of having a `Logger` per module,
or more granular. Having a Logger per function would not be practical,
and expensive, and adding a field to `Event` would lead to a duplication
of `Layout` (as e.g. `LayoutDate` would need to account for it).

The most practical way for the user seems to be to provide the ability
to substitute the Logger name for the name of the function the calls
originate from, which this implements.