Split propagateLevel from propagateValue
propagateValue is used to set an arbitrary field in a Logger to an arbitrary value,
recursively. While convenient, it falls short as soon as the field has some logic
to it. While this can be worked around by exposing specific setters that handle
such logic, we want to move away from simple boolean fields towards
an enum of options, which would mean changing propagateValue - However, as it's
currently used for both the Level and the boolean collect_stats, we first need
to split those two usages.